requirements.txt
langchain==0.3.7
Pillow==11.0.0
streamlit==1.39.0
streamlit_image_select==0.6.0
